Renamed header file to ia_nvml.hpp
